+++ src/adapters/react/index.ts
@@ -2,11 +2,11 @@ import { useState, useEffect } from 'react'
 
 import { ObservableStore } from '../..'
 
-export interface ReactStoreHook {
+export interface ReactObservableHook {
   <S>(store: ObservableStore<S>): S
 }
 
-export const useStore: ReactStoreHook = store => {
+export const useObservable: ReactObservableHook = store => {
   let [, forceUpdate] = useState({})
 
   useEffect(() => store.observe(forceUpdate), [store])