+++ src/index.ts
@@ -1,1 +1,4 @@
+export * from './lazyYobta'
 export * from './observableYobta'
+export * from './replicatedYobta'
+export * from './storageYobta'