+++ src/observableYobta/index.ts
@@ -1,6 +1,7 @@
 import { isEmptyArray } from '../_internal/isEmptyArray'
 import { isFunction } from '../_internal/isFunction'
 
+// #region Types
 interface VoidFunction {
   (): void
 }
@@ -9,48 +10,79 @@ export interface Observer<S> {
   (state: S, ...args: any[]): void
 }
 
-export type Action<S> = S | ((last: S) => S)
-
 export interface ObservableStore<S> {
-  last(): Readonly<S>
-  next(action: Action<S>, ...args: any[]): void
+  last: StateGetter<S>
+  next: StateSetter<S>
   observe(observer: Observer<S>): VoidFunction
 }
 
-interface StoreFactoryOptions {
-  onStart?: VoidFunction
-  onStop?: VoidFunction
+type Action<S> = S | ((last: S) => S)
+
+interface StateGetter<S> {
+  (): S
 }
-interface StoreFactory {
-  <S>(initialState: S, options?: StoreFactoryOptions): ObservableStore<S>
+interface StateSetter<S> {
+  (action: Action<S>, ...args: any[]): void
 }
 
-export const observableYobta: StoreFactory = (initialState, options) => {
-  let { onStart, onStop } = options || {}
+type EventType = 'START' | 'STOP' | 'NEXT'
+
+export interface StoreFactoryEvent<S> {
+  initialState: S
+  last: StateGetter<S>
+  next: StateSetter<S>
+  type: EventType
+}
+export interface StoreFactoryListener<S> {
+  (event: StoreFactoryEvent<S>): void
+}
+
+interface StoreFactory {
+  <S>(
+    initialState: S,
+    ...listeners: StoreFactoryListener<S>[]
+  ): ObservableStore<S>
+}
+// #endregion
+
+export const observableYobta: StoreFactory = <S>(
+  initialState: S,
+  ...listeners: StoreFactoryListener<S>[]
+) => {
   let observers: Observer<any>[] = []
   let state = initialState
 
+  let last: StateGetter<S> = () => state
+  let next = (action: any, ...args: any[]): void => {
+    state = isFunction(action)
+      ? // @ts-ignore
+        action(state)
+      : action
+    observers.forEach(observe => {
+      observe(state, ...args)
+    })
+    emit('NEXT')
+  }
+
+  let emit = (type: EventType): void => {
+    listeners.forEach(send => {
+      send({ initialState, type, last, next })
+    })
+  }
+
   return {
-    last: () => state,
-    next: (action, ...args) => {
-      state = isFunction(action)
-        ? // @ts-ignore
-          action(state)
-        : action
-      observers.forEach(observe => {
-        observe(state, ...args)
-      })
-    },
+    last,
+    next,
     observe: observer => {
-      if (onStart && isEmptyArray(observers)) {
-        onStart()
+      if (isEmptyArray(observers)) {
+        emit('START')
       }
       observers.push(observer)
       return () => {
         let index = observers.indexOf(observer)
         observers.splice(index, 1)
-        if (onStop && isEmptyArray(observers)) {
-          onStop()
+        if (isEmptyArray(observers)) {
+          emit('STOP')
         }
       }
     },