+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Change Log
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.0.11
+* Replace observable config with listeners
+* Implement lazy listener
+* Rename useStore to useObservable
+* Replication plugin
+* Storage plugin
 ## 0.0.10
-* onStart and onStop store options
+* onStart and onStop store optionss
 ## 0.0.9
 * Implement lazyness as an option of observable store
 ## 0.0.8